refactor(checkpoint): move artifact snapshot interface to run types and drop singular field
Relocate the ArtifactSnapshot interface from infra/checkpoint/types.ts
to infra/run/types.ts where its only remaining consumers live (resume
flows). Drop the singular artifactSnapshot field from the checkpoint
webhook contract and its response body — all guest-agents now emit the
plural artifactSnapshots map. Inline the legacy fallback branch in
createCheckpoint that accepted the deprecated singular field, and add
.strict() to the zod body schema so future typos or stale clients fail
fast.

Refs #10743
Parent epic #10740
