+++ turbo/apps/web/src/lib/infra/checkpoint/checkpoint-service.ts
@@ -17,32 +17,6 @@ import type {
 const log = logger("checkpoint");
 
 /**
- * Resolve the artifact snapshot map from a checkpoint request. During the
- * deployment window the guest-agent may still send the legacy singleton
- * `artifactSnapshot`; fold it into the map when the map is missing/empty so
- * no data is lost in mixed old-guest/new-server states.
- */
-function resolveArtifactSnapshots(request: CheckpointRequest): {
-  map: Record<string, string> | null;
-  primaryArtifactName: string | undefined;
-} {
-  const rawMap = request.artifactSnapshots ?? null;
-  const hasMap = rawMap !== null && Object.keys(rawMap).length > 0;
-  if (hasMap) {
-    const primaryArtifactName = Object.keys(rawMap)[0];
-    return { map: rawMap, primaryArtifactName };
-  }
-  const legacy = request.artifactSnapshot;
-  if (legacy) {
-    return {
-      map: { [legacy.artifactName]: legacy.artifactVersion },
-      primaryArtifactName: legacy.artifactName,
-    };
-  }
-  return { map: null, primaryArtifactName: undefined };
-}
-
-/**
  * Create a checkpoint for an agent run
  *
  * @param request Checkpoint request data from webhook
@@ -160,11 +134,10 @@ export async function createCheckpoint(
       }
     : null;
 
-  // Consolidate artifact snapshots. The guest-agent's authoritative payload
-  // is artifactSnapshots (name -> version map); during rollout a legacy
-  // singleton artifactSnapshot is folded into the map if present.
-  const { map: artifactSnapshotsMap, primaryArtifactName } =
-    resolveArtifactSnapshots(request);
+  const rawMap = request.artifactSnapshots ?? null;
+  const hasMap = rawMap !== null && Object.keys(rawMap).length > 0;
+  const artifactSnapshotsMap = hasMap ? rawMap : null;
+  const primaryArtifactName = hasMap ? Object.keys(rawMap)[0] : undefined;
 
   const snapshotFields = {
     conversationId: conversation.id,
