+++ turbo/apps/web/src/lib/infra/checkpoint/types.ts
@@ -15,15 +15,6 @@ export interface AgentComposeSnapshot {
 }
 
 /**
- * Artifact snapshot for VAS managed artifacts
- * Fields align with CLI parameters --artifact-name and --artifact-version
- */
-export interface ArtifactSnapshot {
-  artifactName: string;
-  artifactVersion: string;
-}
-
-/**
  * Volume versions snapshot for checkpoint
  * Stores resolved volume versions at checkpoint time for exact reproducibility
  */
@@ -46,10 +37,9 @@ export interface CheckpointRequest {
   cliAgentType: string;
   cliAgentSessionId: string;
   cliAgentSessionHistoryHash: string;
-  artifactSnapshot?: ArtifactSnapshot;
   // Multi-artifact snapshot map: artifactName -> versionId. Emitted
-  // unconditionally by the guest-agent during the multi-mount rollout;
-  // may be empty or missing when the guest snapshotted nothing.
+  // unconditionally by the guest-agent; may be empty or missing when the
+  // guest snapshotted nothing.
   artifactSnapshots?: Record<string, string>;
   volumeVersionsSnapshot?: VolumeVersionsSnapshot;
 }
@@ -61,7 +51,6 @@ export interface CheckpointResponse {
   checkpointId: string;
   agentSessionId: string;
   conversationId: string;
-  artifact?: ArtifactSnapshot;
   artifacts?: Record<string, string>;
   volumes?: Record<string, string>;
 }
