+++ turbo/apps/web/src/lib/infra/run/executors/types.ts
@@ -1,6 +1,5 @@
 import type { StorageManifest } from "../../../infra/storage/types";
-import type { ResumeSession } from "../types";
-import type { ArtifactSnapshot } from "../../checkpoint/types";
+import type { ResumeSession, ArtifactSnapshot } from "../types";
 import type { Firewalls, NetworkPolicies } from "@vm0/core";
 
 /**
